refactor: improve: move contributors config in `creator.ts`to `metadata.ts` for better UX/DX #55
