+++ .vitepress/theme/components/HomePage.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { VPTeamMembers } from 'vitepress/theme'
-import { creators } from '../../creators'
-import { siteName } from '../../../metadata'
+import { creators, siteName } from '../../../metadata'
 </script>
 
 <template>
