+++ scripts/types/metadata.d.ts
@@ -38,3 +38,21 @@ export interface DocsTagsAlias {
   name: string
   alias: string[]
 }
+
+export interface SocialEntry {
+  type: 'github' | 'twitter' | 'email'
+  icon: string
+  link: string
+}
+
+export interface Creator {
+  avatar: string
+  name: string
+  username?: string
+  title?: string
+  org?: string
+  desc?: string
+  links?: SocialEntry[]
+  nameAliases?: string[]
+  emailAliases?: string[]
+}
