+++ scripts/utils.ts
@@ -1,0 +1,1 @@
+export const getAvatarUrlByGithubName = (name: string) => `https://github.com/${name}.png`
