+++ vite.config.ts
@@ -7,8 +7,7 @@ import { GitChangelog, GitChangelogMarkdownSection } from '@nolebase/vitepress-p
 import { PageProperties, PagePropertiesMarkdownSection } from '@nolebase/vitepress-plugin-page-properties/vite'
 import { ThumbnailHashImages } from '@nolebase/vitepress-plugin-thumbnail-hash/vite'
 
-import { githubRepoLink } from './metadata'
-import { creators } from '.vitepress/creators'
+import { creators, githubRepoLink } from './metadata'
 
 export default defineConfig(async () => {
   return {
